Added Chinese simplified translation
Known issue: "Scenario" not translated.
Unsure how to solve:
in Main Window column, the entry for "FileMenuItem", "EditMenuItem", "ViewMenuItem", "LanguageMenuItem", and "InfoMenuItem", (as an example I refer to FileMenuItem but it also applies to others that I mentioned), what would show up in the main window should be "文件（F）" without quotation marks, but it is using "F" as the shortcut the same as using English translation. I hope what I did does what I said, but I need confirmation about it.